Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
boto3
python-multipart
